diag: transport: Add support for USB connection to host PC
Add the option to connect to a host PC connected via a USB cable.
Configure the USB via the following command line option:
-g <gadget device name[#serial number]>
The device name is typically /dev/ffs_diag.
Note that this is based on a user space erected driver using configfs and
requires a gadget and diag function file system nodes to be setup a priori.

Change-Id: I4de1d58890602c804dbe89de199064b235d4d914
Signed-off-by: Eyal Ilsar <eilsar@codeaurora.org>