diag: peripheral: plugin: Add support for peripheral extension libraries
In order to support special command handling the peripheral was extended
to be also a dynamically loaded plugin that is implemented in a shared
object library.
The interface between the plugin and diag allows for registration of
command ranges with special handlers and a data channel like input
for the response packets.

Change-Id: I656b50f40914af690c3b9ad0908aac53b77d3eb1
Signed-off-by: Eyal Ilsar <eilsar@codeaurora.org>diag: peripheral: plugin: Add support for peripheral extension libraries

In order to support special command handling the peripheral was extended
to be also a dynamically loaded plugin that is implemented in a shared
object library.
The interface between the plugin and diag allows for registration of
command ranges with special handlers and a data channel like input
for the response packets.

Change-Id: I656b50f40914af690c3b9ad0908aac53b77d3eb1
Signed-off-by: Eyal Ilsar <eilsar@codeaurora.org>