diag: peripheral: plugin: Add FTM_BT plugin
This plugin intercepts and handles Factory Test Mode Bluetooth related
commands and forwards them to the BT subsystem via HCI commands over the
APPS_RIVA_BT_CMD channel.

For proper operation please unload the btqcomsmd module as it holds this
channel open when active.

To include this plugin add the following to the diag command line:
-p /usr/lib/ftm_bt.so

Change-Id: I5c6528b6540a088019534195da04e97dc8952687
Signed-off-by: Eyal Ilsar <eilsar@codeaurora.org>