diag: Facilitate packet encoding and enqueuing via a utility function
Moved the packet encoding and enqueuing code from main code to utility and
use it throughout the different layers.
On the way revised packet en/decoding logic assuming:
a) incoming packets from clients are always encoded and decoded upon
reception.
b) outgoing packets to clients are always encoded.
c) forwarded packets to peripherals are NOT required to be encoded if apps encoding
encoding feature is set.
d) response packets from peripherals are NOT encoded if the apps encoding
feature is set.

Change-Id: I35f8ac94fafff9569e64fd9443b2342188b3ee53
Signed-off-by: Eyal Ilsar <eilsar@codeaurora.org>